backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.26.0
anthropic>=0.18.0
pydantic>=2.5.0
supabase>=2.0.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
google-api-python-client>=2.100.0
google-auth>=2.25.0
notion-client>=2.2.0
eval-type-backport>=0.2.0
